+++ src/minecraft/co/uk/flansmods/common/EntityMG.java
@@ -212,7 +212,7 @@ public class EntityMG extends Entity implements IEntityAdditionalSpawnData
 	}
 
 	@Override
-	public boolean attackEntityFrom(DamageSource damagesource, int i)
+	public boolean attackEntityFrom(DamageSource damagesource, float i)
 	{
 		if (damagesource.damageType.equals("player"))
 		{
